# openapi.yaml
openapi: 3.0.3
info:
  title: Grant Application Portal API
  version: 1.0.0
  description: >
    Backend for the grant application portal: applicants browse grants, fill in
    a form generated from per-grant requirements, upload documents and submit;
    staff define grants, review applications and export the results.
servers:
  - url: http://localhost:4000
    description: Local development
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, description: 64-bit id serialized as a string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [applicant, admin, reviewer] }
        organization: { type: string, nullable: true }
...
    Review:
      type: object
      properties:
        id: { type: string }
        applicationId: { type: string }
        reviewerId: { type: string }
        score: { type: number, nullable: true }
        notes: { type: string, nullable: true }
        recommendation:
          type: string
          nullable: true
          enum: [approve, reject, needs_more_info]
        reviewer: { $ref: '#/components/schemas/User' }
...
paths:
...
  /reviews:
    get:
      tags: [Reviews]
      summary: List reviews (staff)
      parameters:
        - { in: query, name: application_id, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Review' }
    post:
      tags: [Reviews]
      summary: Create or update your review of an application (staff)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [application_id]
              properties:
                application_id: { type: string }
                score: { type: number, minimum: 0, maximum: 100, nullable: true }
                notes: { type: string, nullable: true }
                recommendation:
                  type: string
                  nullable: true
                  enum: [approve, reject, needs_more_info]
      responses:
        '201':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Review' }
  /reviews/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    delete:
      tags: [Reviews]
      summary: Remove your own review (admins may remove any)
      responses:
        '204': { description: Deleted }
...
  /users:
    get:
      tags: [Users]
      summary: List users (admin)
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: role, schema: { type: string, enum: [applicant, admin, reviewer] } }
      responses:
        '200': { description: OK }
    post:
      tags: [Users]
      summary: Create a user with a role (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string, minLength: 8 }
                role: { type: string, enum: [applicant, admin, reviewer] }
                organization: { type: string }
      responses:
        '201': { description: Created }
  /users/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      tags: [Users]
      summary: Update a user's role, status, name or password (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                role: { type: string, enum: [applicant, admin, reviewer] }
                organization: { type: string, nullable: true }
                is_active: { type: boolean }
                password: { type: string, minLength: 8 }
      responses:
        '200': { description: Updated }
        '400': { description: You cannot change your own role or deactivate yourself }
...